+++ packages/office-ui-fabric-react/src/components/Dropdown/Dropdown.types.ts
@@ -21,7 +21,8 @@ export interface IDropdownProps extends ISelectableDroppableTextProps<IDropdown,
   placeHolder?: string;
 
   /**
-   * Options for the dropdown.
+   * Options for the dropdown. If using `defaultSelectedKey` or `defaultSelectedKeys`, options must be
+   * pure for correct behavior.
    */
   options: IDropdownOption[];