+++ packages/styling/src/interfaces/ITheme.ts
@@ -1,7 +1,6 @@
 import { IPalette } from './IPalette';
 import { IFontStyles } from './IFontStyles';
 import { ISemanticColors } from './ISemanticColors';
-import { ITypography, IPartialTypography } from './ITypography';
 import { ISpacing } from './ISpacing';
 import { IEffects } from './IEffects';
 
@@ -31,14 +30,6 @@ export interface IScheme {
 
   /**
    * @internal
-   * The typography property is still in an experimental phase. The intent is to have it
-   * eventually replace IFontStyles in a future release, but it is still undergoing review.
-   * Avoid using it until it is finalized.
-   */
-  typography: ITypography;
-
-  /**
-   * @internal
    * The spacing property is still in an experimental phase. The intent is to have it
    * be used for padding and margin sizes in a future release, but it is still undergoing review.
    * Avoid using it until it is finalized.
@@ -63,5 +54,4 @@ export type IPartialTheme = {
     ITheme,
     'palette' | 'fonts' | 'semanticColors' | 'isInverted' | 'disableGlobalClassNames' | 'spacing' | 'schemes' | 'effects'
   >]?: Partial<ITheme[P]>
-} &
-  { [P in keyof Pick<ITheme, 'typography'>]?: IPartialTypography };
+};